Suggest creating landing page
Love this idea and the design looks awesome! One suggestion: on the live version of the app, when I went to https://log-my-pet.vercel.app, it first showed me this page before redirecting me to the login page. Suggest creating a landing page with signup/login links so the user doesn't need to land on this homepage first.