General comments
Awesome work!! Your app is really cool, visually pleasing and fun to use. And your code is extremely neat and organized.

Personally, I'd appreciate some styling when users hover over any kind of action button, but maybe that's just personal preference. Also, there are a ton of records on the page, so it might be nice to add pagination as well as filtering, sorting, etc.

Anyway, it's all really well done - congrats on a very successful app!